+++ src/components/auth/register/RegisterCompeleteModal/style.ts
