+++ src/components/auth/verify/DuplicateEmailModal/style.ts
